apps/invoices/serializers.py
from rest_framework import serializers

from apps.invoices.models import Invoice

class InvoiceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Invoice
        fields = '__all__'